api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
python-dotenv>=1.0.0
numpy
rasterio
GDAL
opencv-python
scikit-image
